requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.7
dj-database-url==2.2.0
Django==4.1
django-cors-headers==4.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
h11==0.14.0
nodejs==0.1.1
npm==0.1.1
optional-django==0.1.0
packaging==24.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
whitenoise==6.8.2
